Move messages in client to its own folder
