+++ client/core/src/com/utilities/messages/Message.java
@@ -1,4 +1,4 @@
-package com.utilities;
+package com.utilities.messages;
 
 public abstract class Message {
     public int roomCode;
