+++ client/core/src/com/utilities/messages/SomeResponse.java
@@ -1,4 +1,4 @@
-package com.utilities;
+package com.utilities.messages;
 
 public class SomeResponse extends Message {
     public String text;
